moved stagefactory to lucille-core util folder
