+++ lucille-core/src/test/java/com/kmwllc/lucille/stage/RenameFieldsTest.java
@@ -3,6 +3,7 @@ package com.kmwllc.lucille.stage;
 import com.kmwllc.lucille.core.Document;
 import com.kmwllc.lucille.core.Stage;
 import com.kmwllc.lucille.core.StageException;
+import com.kmwllc.lucille.util.StageFactory;
 import org.junit.Test;
 
 import java.util.Set;
